FEAT: Implement consolidated ClientOrchestrator for Phase 4
- Created acme/client_orchestrator.py (~495 lines)
- Extracted client lifecycle management from backend.py
- Extracted computation orchestration from backend.py
- Extracted progress monitoring and error handling
- Enhanced error handling organization with dedicated methods
- Backend.py reduced from 747 → 445 lines (40% reduction)
- Maintained func_wrapper as static method in backend.py
- Preserved all SLURM error analysis functionality
- Improved code organization while maintaining backward compatibility

Key improvements:
- Unified complexity in single ClientOrchestrator class
- Clear separation of concerns (client lifecycle vs orchestration)
- Enhanced error handling modularity
- Consistent patterns with Phase 1-3 components
- Config-driven initialization leveraging ACMEConfig
